New locking stategy for NetCDFDataProxy.
