+++ lib/iris/fileformats/netcdf/_thread_safe.py
@@ -10,12 +10,13 @@ Intention is that no other Iris module should import the netCDF module.
 
 """
 from abc import ABC
-import threading
 import typing
 
+from dask.utils import SerializableLock
 import netCDF4
+import numpy as np
 
-_GLOBAL_NETCDF4_LOCK = threading.Lock()
+_GLOBAL_NETCDF4_LOCK = SerializableLock()
 
 # Doesn't need thread protection, but this allows all netCDF4 refs to be
 #  replaced with thread_safe refs.
@@ -291,3 +292,51 @@ class DatasetContainer(GroupContainer):
         with _GLOBAL_NETCDF4_LOCK:
             instance = cls.CONTAINED_CLASS.fromcdl(*args, **kwargs)
         return cls._from_existing(instance)
+
+
+class NetCDFDataProxy:
+    """A reference to the data payload of a single NetCDF file variable."""
+
+    required_lock = _GLOBAL_NETCDF4_LOCK
+
+    __slots__ = ("shape", "dtype", "path", "variable_name", "fill_value")
+
+    def __init__(self, shape, dtype, path, variable_name, fill_value):
+        self.shape = shape
+        self.dtype = dtype
+        self.path = path
+        self.variable_name = variable_name
+        self.fill_value = fill_value
+
+    @property
+    def ndim(self):
+        return len(self.shape)
+
+    def __getitem__(self, keys):
+        # Dask will call this many times during reading, and also provides
+        #  dedicated API for lock acquisition, so we use that instead of using
+        #  a DatasetContainer ( see iris._lazy_data.as_lazy_data() ).
+        assert self.required_lock.locked()
+        dataset = netCDF4.Dataset(self.path)
+        try:
+            variable = dataset.variables[self.variable_name]
+            # Get the NetCDF variable data and slice.
+            var = variable[keys]
+        finally:
+            dataset.close()
+        return np.asanyarray(var)
+
+    def __repr__(self):
+        fmt = (
+            "<{self.__class__.__name__} shape={self.shape}"
+            " dtype={self.dtype!r} path={self.path!r}"
+            " variable_name={self.variable_name!r}>"
+        )
+        return fmt.format(self=self)
+
+    def __getstate__(self):
+        return {attr: getattr(self, attr) for attr in self.__slots__}
+
+    def __setstate__(self, state):
+        for key, value in state.items():
+            setattr(self, key, value)
